Chore: move RepoCommand match to handle_repo
