+++ src/cli_parse/handle_commands/mod.rs
@@ -1,2 +1,3 @@
 pub mod handle_issue;
 pub mod handle_comment;
+pub mod handle_repo;
